Create classification model
Beaches alongside Chicago's Lake Michigan are equipped with weather sensors implemented by The Chicago Park District. Hourly Measurements are captured by these sensors and broader insights can be taken out by evaluating this data and implementing various machine learning algorithm. Text classification one of the widely used natural language processing. We have readings in dataset from the sensors. Our main focus is on total rain and all the dependent variables on our target variable. At first I identifies the target variable by understanding the dataset and identified what best outcome can be used for prediction and analysis purpose. There was a column named Total Rain which was my target variable. At the initial stage I cleaned the data by replacing all null values by mean and median of their respective columns. I also eliminated the columns which were not of concerning after selecting the target variable. But there was a problem because the data of target variable was continuous .Also I needed classify the column as in sub categories as very low(VL),low(L),moderate(M),very high(VH) for analysis. To deal with this continuous data I classified the data of the column into four sub categories by putting them in bins and shifted these sub categories to a new column name rain level by implementing bucket .Then I dropped the Total rain column to avoid data redundancy. Next step was label encoder in which we assign labels. Then we deal with de pendent and independent variable by inserting them in X and Y label .Then I did scaling to standardize the data. Then comes the part of testing and training where 30% is test data.
Then comes the part of implementing various models. At first I have implemented decision tree classifier and obtained the confusion matrix and accuracy. Then the prediction is done on the basis of decision tree classifier model. The accuracy was 100 percent which is not possible while doing prediction. This was the indication that the training data was over fit for this model and tis model is unsuitable of the dataset. Then I went forward with applying another model to the dataset. I applied support vector machine (SVM) and predicted the model. After words I printed the confusion matrix and accuracy. The accuracy gained by SVM was optimal as compared to decision tree classifier .And we move forward with SVM model. In conclusion we can say SVM has given a better prediction that which type of rainfall weather its very low or low or moderate or very high. And using this prediction in future we can classify locations on the basis of rain falls and predict the rain fall for upcoming years. I have encountered lots of challenges while performing the coding part. First off all finding the dataset was a task because the model implementation and source of dataset was limited. Also making a decision best model selection was also a crucial part in all the process.