Create Create regression model
Regression is the process of finding the relationship between target variable and a combination of other independent variables, where the required data type of dependent variable is continuous in nature. It is also useful in various fields like computer science, economy, and social science and so on. The implementation of regression rises day to day where we need to deal with large amount of data. The dataset used for applying regression analysis contains estimating records of daily streamflow at ungagged locations in the southeast United States. This dataset contains 9 attributes and 293 instances. In U.S. Multiple variables hold the values of gauges which are considered as independent variables where. The 90th percentile of annual maximum streamflow is provided as an example target variable for 293 stream gages. Here we had an information in the form of dataset regarding water resources for accurate measurement of streamflow .Most crucial part is installing all the required libraries for the whole process. Then comes the identification of target variable, and here it’s the column named “MAR_PPT7100_CM”. To perform regression model we need continuous dependent variable so in our process this requirement is fulfilled where the response variable is “MAR_PPT7100_CM”, which is continuous in nature. Then we imported data and started to clean it for the betterment of prediction. Initially, the missing values in the data are checked and accordingly, there were no presence of null values. Furthermore, scaling of independent variables is been carried out using a Standard Scaler. The total
data is then divided into a training and testing subsets in the ratio of 70% and 30% respectively. Here, the pre-processing and data transformation is successfully completed. After extracting the data in the required form, a Decision Tree Regressor is built at initial stage. The DecisionTreeRegressor() function is imported using the sklearn library. Accordingly, the model is fitted on the training data and tested using testing. Subsequently, the model produced an optimal accuracy of around 89%. Furthermore, a Support Vector Machine regressor is implemented which yielded an accuracy of around 58%, which is comparatively very low as of decision tree. These model were implemented without any hyper parameter tuning which yielded their highest accuracy. As the above mentioned method were non-parametric, which makes no assumptions on the given data. Thus, on the other hand a parametric algorithm namely, a Linear regression is also designed. In our case the linear model was built without any assumption satisfaction and surprisingly, the model produced an excellent accuracy of 85% which was higher than SVM. Here we are dealing with multiple linear regression. We can conclude with how the models are helpful in getting a good overview of streamflow which will help in predicting the quality and capability of water resources in the specific location.